Added Shipments

+++ ResataurantClient/Services/CartService.cs
@@ -12,29 +12,30 @@ namespace ResataurantClient.Services
             _httpClient = httpClient;
         }
 
-        public async Task<Cart> GetCart()
+        public async Task<Cart> GetCart(int userId)
         {
-            return await _httpClient.GetFromJsonAsync<Cart>("api/cart");
+            return await _httpClient.GetFromJsonAsync<Cart>($"api/cart/{userId}");
         }
 
-        public async Task AddToCart(CartItem item)
+        public async Task AddToCart(int userId, CartItem item)
         {
-            await _httpClient.PostAsJsonAsync("api/cart", item);
+            await _httpClient.PostAsJsonAsync($"api/cart/{userId}", item);
         }
 
-        public async Task RemoveFromCart(int productId)
+        public async Task RemoveFromCart(int userId, int productId)
         {
-            await _httpClient.DeleteAsync($"api/cart/{productId}");
+            await _httpClient.DeleteAsync($"api/cart/{userId}/{productId}");
         }
 
-        public async Task ClearCart()
+        public async Task ClearCart(int userId)
         {
-            await _httpClient.DeleteAsync("api/cart");
+            await _httpClient.DeleteAsync($"api/cart/{userId}/clear");
         }
-        public async Task UpdateQuantity(int productId, int newQuantity)
+
+        public async Task UpdateQuantity(int userId, int productId, int newQuantity)
         {
             var updateQuantityRequest = new { ProductId = productId, NewQuantity = newQuantity };
-            await _httpClient.PutAsJsonAsync("api/cart/update-quantity", updateQuantityRequest);
+            await _httpClient.PutAsJsonAsync($"api/cart/{userId}/update-quantity", updateQuantityRequest);
         }
     }
